dom/tests/mochitest/bugs/test_bug648465.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=648465
-->
<head>
  <title>Test for Bug 648465</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank"
  href="https://bugzilla.mozilla.org/show_bug.cgi?id=648465">
  Mozilla Bug 648465</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 648465 **/
var w = window.open('about:blank');
try {
  is(w.fullScreen, false, "can access non-allAccess properties of w");
} finally {
  w.close();
}

</script>
</pre>
</body>
</html>
